examples/google_adk/mcp_brave_search_agent/pyproject.toml
[project]
name = "adk-mcp-brave-search-agent-example"
version = "0.1.0"
description = "MCP Brave Search agent example"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "mcp",
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "google-genai>=1.9.0",
    "google-adk>=1.0.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-sdk = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
